update tests to run faster
