+++ .gitignore
@@ -7,6 +7,7 @@ current_bird_release.csv
 *.pl
 *.pt
 tests/__pycache__
+tests/data/*
 *.wp*
 lightning_logs/*
 *.prof
