+++ tests/test_IoU.py
@@ -10,9 +10,7 @@ import shapely
 import geopandas as gpd
 import pandas as pd
 
-def test_compute_IoU(download_release, tmpdir):
-    m = main.deepforest()
-    m.use_release(check_release=False)
+def test_compute_IoU(m, download_release, tmpdir):
     csv_file = get_data("OSBS_029.csv")
     predictions = m.predict_file(csv_file=csv_file, root_dir=os.path.dirname(csv_file))
     ground_truth = pd.read_csv(csv_file)
