+++ tests/test_main.py
@@ -24,41 +24,6 @@ from PIL import Image
 from .conftest import download_release
 
 @pytest.fixture()
-def two_class_m():
-    m = main.deepforest(num_classes=2,label_dict={"Alive":0,"Dead":1})
-    m.config["train"]["csv_file"] = get_data("testfile_multi.csv") 
-    m.config["train"]["root_dir"] = os.path.dirname(get_data("testfile_multi.csv"))
-    m.config["train"]["fast_dev_run"] = True
-    m.config["batch_size"] = 2
-        
-    m.config["validation"]["csv_file"] = get_data("testfile_multi.csv") 
-    m.config["validation"]["root_dir"] = os.path.dirname(get_data("testfile_multi.csv"))
-    m.config["validation"]["val_accuracy_interval"] = 1
-
-    m.create_trainer()
-    
-    return m
-
-@pytest.fixture()
-def m(download_release):
-    m = main.deepforest()
-    m.config["train"]["csv_file"] = get_data("example.csv") 
-    m.config["train"]["root_dir"] = os.path.dirname(get_data("example.csv"))
-    m.config["train"]["fast_dev_run"] = True
-    m.config["batch_size"] = 2
-        
-    m.config["validation"]["csv_file"] = get_data("example.csv") 
-    m.config["validation"]["root_dir"] = os.path.dirname(get_data("example.csv"))
-    m.config["workers"] = 0 
-    m.config["validation"]["val_accuracy_interval"] = 1
-    m.config["train"]["epochs"] = 2
-    
-    m.create_trainer()
-    m.use_release(check_release=False)
-    
-    return m
-
-@pytest.fixture()
 def raster_path():
     return get_data(path='OSBS_029.tif')
